+++ src/apps/relay/dtls_listener.h
@@ -42,6 +42,7 @@
 #include "ns_turn_server.h"
 
 #include <event2/event.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -58,9 +59,10 @@ typedef struct dtls_listener_relay_server_info dtls_listener_relay_server_type;
 void setup_dtls_callbacks(SSL_CTX *ctx);
 #endif
 
-dtls_listener_relay_server_type *create_dtls_listener_server(const char *ifname, const char *local_address, int port,
-                                                             int sock_buf_size, int verbose, ioa_engine_handle e,
-                                                             turn_turnserver *ts, int report_creation,
+dtls_listener_relay_server_type *create_dtls_listener_server(const char *ifname, const char *local_address,
+                                                             uint16_t port, int sock_buf_size, int verbose,
+                                                             ioa_engine_handle e, turn_turnserver *ts,
+                                                             int report_creation,
                                                              ioa_engine_new_connection_event_handler send_socket);
 
 void udp_send_message(dtls_listener_relay_server_type *server, ioa_network_buffer_handle nbh, ioa_addr *dest);
